+++ scripts/ci/pylint_todo.txt
@@ -89,7 +89,7 @@
 ./airflow/sensors/python.py
 ./airflow/contrib/sensors/qubole_sensor.py
 ./airflow/providers/redis/sensors/redis_key.py
-./airflow/contrib/sensors/redis_pub_sub_sensor.py
+./airflow/providers/redis/sensors/redis_pub_sub.py
 ./airflow/providers/amazon/aws/sensors/sagemaker_base.py
 ./airflow/providers/amazon/aws/sensors/sagemaker_endpoint.py
 ./airflow/providers/amazon/aws/sensors/sagemaker_training.py
